got it kinda working
This patch contains changes that get shpool working sufficiently
well enough so that you can `shpool attach` and get a working
shell where you can run `ls` and see sensible results. It is
still quite broken in lots of ways though. The most obvious
way is that there is no prompt visible in the shell you get
when you `shpool attach` (at first I was thinking it was because
PS1 gets clobbered, which it does, but I now think the real issue
is that bash notices that its stdout is not going to a terminal
(as determined by isatty)).

Also, I don't see all the log lines I was expecting to see and
after I exited the shell `shpool list` just hung.

Change-Id: I9a84a239667b9facc9306eddc91e19b0fd261b4b