make some more progress on attach functionality
I've managed to get a prompt to appear, but it appears
in the stderr log of the daemon rather than the shell
session window. I think once I add a little framing
protocol so I can squirt both stdout and stderr data
through the unix socket, we should be good to go.

Another bug that has become apparent is that when a
shell is created, the $HOME variable is not getting
set, which causes many of my setup commands in
.profile and .bashrc to fail, since they reference $HOME
quite a bit. I need to figure out how $HOME is set.

Change-Id: I6badfddd2f72799cc05fce5f2c4405bd9da34926