add ssh extension mode
This patch adds a new POC ssh extension mode which uses shpool to
add session persistance support to native ssh commands. I've gotten
dialing into a new shell session to work, but reattaching still
requires some debugging.

The main idea is to add a couple of new plumbing commands that can
be invoked from the ssh RemoteCommand and LocalCommand config options.
Ideally we would be able to generate a unique identifier for the client
tty from the remote command, but we can't do this because we don't know
anything about the client shell from within the remote shell. To
hack around this, we can use a LocalCommand to generate a unique identifier
based off the local tty number, then do a second ssh invocation to
smuggle this identifier over to the target host.

Change-Id: I8781f96cc539ea97d14ea6b5f518ec81054c642c