disable echo mode for subshells
By default new ptys are set up to echo your commands back to you,
but terminal emulators already do this generally, so you end up
seeing everything double. This patch teaches the daemon to flip
the ECHO flag in the terminal off before spawning a subshell.
I'm pretty sure it even manages to avoid allocating to be fork-safe
as well.

Change-Id: Ib0c7e5678f0bcd957eec2a5590ae1b464055c418