fix race on immediate reconnect
This patch fixes a race that was causing the exit_immediate_drop
test to flake.

The problem was that if a client killed a child shell from within
the shell (by typing `exit` for example) and then immediately dropped
the connection and reconnected. The record for the child shell would
stick around in the session table, the shell would exit, then the
reconnecting attach client would attempt to attach to a dead shell
and immediately get booted since the shell was dead. The fix for this
was to re-check if the shell had exited uppon reconnect.

In the process of doing this, I also realized that checking if the
pty master fd was closed was not the best way to see if the child shell
was still around, since it leaves a bunch of zombie processes around,
and is also not as reliable as just reaping the children correctly.
We now spin up a dedicated reaper thread when a new child shell is
launched so we can immediately detect when it exits and reap the
exit code.

Change-Id: Iae175982c0eaa294900f166eacbdb393b03ba04b