set the correct pty size on attach
This patch teaches shpool to forward the local tty size
to the pty that the shell gets opened in on attach. This
is required for cureses based/full screen applications
like less or vim to work correctly. We don't currently
handle SIGWINCH signals or forward them to the shell
process, but we will need to do that to correctly handle
terminal windows resizing.

Change-Id: Id4490b08a74aecdb9c92618e3b126dd357231d0e