add detach subcommand
This patch adds a new subcommand to allow you to gracefully
detach a running session from shpool. When run from inside
the session in question, shpool can infer what you mean
if you leave off the list of session names to detach
from. Piping the output of `shpool list` to `shpool detach`
is left as an exercise for the reader.

Change-Id: I6661496e903f3bab2d956cc1c2c25bb79abe608e