clean up TODO file
It turns out SIGWINCH is a magical fix for curses app support.
Everything seems to Just Work (tm) when you reattach to a
full screen app (though sometimes you need to jiggle your terminal
size to kick it, so we might need to sprinkle a few more hacks in
place to get it totally working).

Change-Id: Id4e5a73c2812bcc9031a1a3f4533690bd21386cb