use symlink to dynamically point to SSH_AUTH_SOCK
This patch switches the approach that shpool takes to SSH_AUTH_SOCK
in order to support hardware security keys. Previously, we would
just inject the env var into the shpool shell session, but this
runs into trouble because the SSH_AUTH_SOCK value changes
when a connection drops and someone shells back in. Unfortunately,
it is not really possible to change environment variables in another
process dynamically without resorting to stuff like LD_PRELOAD hacks
or using a debugger to rummage around in the processes' symbol table.

What we can do, is set SSH_AUTH_SOCK to point to a symlink when we
launch the shell session, then just update that symlink to point to
the right place on reattach. This patch does that.

I've moved the default location of the shpool socket from
$XDG_RUNTIME_DIR/shpool.socket to $XDG_RUNTIME_DIR/shpool/shpool.socket
in this patch because I need to put these new SSH_AUTH_SOCK links
somewhere and I want to keep everything shpool related namespaced
in a single directory. This may cause some user disruption.

Change-Id: I91ddf6edb3ef010d9e61fcf131f16532c27c4277