fix reattach redraw bug for emacs
This patch tweaks our SIGWINCH reattach trick so that
it works for emacs. For some reason, the way we were jiggling
the pty size before was not working, but this new approach
works for emacs. I had to use the RPC resize mechanism rather
than just directly resizing the pty, which is pretty weird,
but at least it is working.

Change-Id: I263e91b0f40196c775c21fd45824d2b9c3f9dd76