add keybinding engine
This patch adds a new engine for scanning over the input
stream looking for keybindings. To do so, we allow users
to specify keybindings using standardish Ctrl-x style
syntax. The engine supports both chords and sequences,
though chord support is quite limited. For now, we only
allow Ctrl followed by a single symbol as a chord,
since these chords generate a single byte.
The the engine is designed to be able to handle
chords which generate multiple different codes
though.

The core of the matching algorithm is a pair of tries. The
first line trie maps byte sequences to chords (for now those
sequences are always of length 1, but we should be able to
easily extend this). The second line trie maps chord sequences
to actions. We basically just keep pushing bytes into the first
line trie, and any time it emits a chord, we push that chord
into the second line trie, emitting an action whenever the
second line trie gets a match.

Right now, we don't actually perform any actions, just log
them. Also, we are doing the matching all in the client->shell
thread, and I'm not sure this is the best approach in terms of
performance (though I'm also not sure that a memcpy to a background
thread will be much better, simd should help a little I guess).

Change-Id: I99821a05d08285524ad72ba3cf2bf332ea3a4132