add -f flag to attach subcommand
In some cases, such as when you're running an automatic
reconnect loop from the client, you don't want to be too
polite about it if someone else is already attached to
a session you are trying to attach to. This patch adds
a new flag to the `attach` subcommand which will automatically
detach a session if it has another tty attached.

Change-Id: I4864a940f625944d2cbdc31b7715ed02cf76ad5d