vendor dependencies
This patch vendors all our dependencies so that we can perform
builds without talking to the internet. I'm still not all that
bullish on the potential to do kokoro builds in completely
isolated VMs, since I'm not sure how to ensure that a good
rust toolchain is installed. Still, it is a bit of progress.

Change-Id: Id402d42682777419c332c1cfc83eda8e5a4b8663