document how to measure shpool latency
This patch documents how to measure shpool latency
in HACKING.md, and records are current (pretty terrible)
numbers. I'm going to try switching to syncronous IO
to see if this improves things.

I've also tried directly instrumenting the binary to measure
latency, and it found simillarly bad numbers (actually a bit
worse), but that code is kinda gross and I don't know that
I want to check it in when an e2e tool like sshping works.

For some reason even though I only made doc changes in this
patch, I ran into CI issues, so this change includes some CI
fixes as well. The vendor churn is because I needed to add
the `rand` crate in order to support landing test logs in
the "Artifacts" tab in kokoro spounge links.

Change-Id: I8b313b2c275c5d05f1cd461af27595d3e1b39d62