deflake tests in CI
This patch deflakes the shpool tests, which are failing
intermittantly in kokoro. There were a few different
problems.

The problem for exit_immediate_drop was that input
chunks sometimes get batched up which messes with the
regex matchers. Solving this involved adding a couple of
new events, which in turn created new problems by flooding
the event socket buffer and causing output_flood to lock up,
so I had to add a flag to the daemon test proc constructor
to allow control over creating an event subscriber.

The problem for multiple_mixed was we were waiting on the
right events, but out of order.

The vendor churn is because I added the `ntest` crate
in order to be able to set timeouts on tests. Somehow,
that isn't a first party cargo feature. Several of my
kokoro test runs got stuck with tests running forever,
which meant I never got to see the logs.

Change-Id: I7a2e4f06cb389b9a84241a2ab70fdeb144e60a35