add benz release config
This patch adds a benz release config and release
script. In order to generate the .deb file, it
vendors the `cargo deb` tool.

I verified that this script works by tweaking
the presubmit.cfg file to point to the benz build
script instead of the normal test script, so I
know it works in a kokoro environment using the
same image benz uses, but I'm not 100% sure it
will work in actual benz. Hopefully, it will only
take minor tweaks to get it working in real benz.

Change-Id: Ic1c77961fe1edaa7238cd1acf57f174d4598fe7d