attempt to fix benz build
Benz can now pick up the shpool repo and attempt to build it,
but it is failing when the presence checks for various debian/*
files fail. I think this is not needed since we are using `cargo deb`
rather than the normal debian tooling for building our package.

Here is an example failure:
https://fusion2.corp.google.com/invocations/fe85a3d0-8f9a-42a6-ba3d-47e881390621/targets/goobuntu-benz%2Fbenz%2Fe7e3c81e-673c-4c5a-b9ee-6329bb8ece16;config=default/log

The ongoing campaign to deflake the tests in kokoro presubmits
has also popped up in this change, in the form of making failed
resize RPCs non-fatal to a shell session.

Change-Id: I104e780a724492b294d227182471516f35dfc609