hook up detach action
This patch hooks up the detach keybinding action to make it actually
detach. The default detach keybinding is "Ctrl-Space Ctrl-q", but a
custom keybinding can be configured in the config.toml file.

Before announcing this on the mailing list and closing the bug,
I need to make sure that the install script still works with
new vender changes, document it (including limitations of the keybinding
engine), and add a default location for the shpool config, so people
don't need to edit their systemd files in order to customize their keybinding.

Change-Id: I5fa0605b49cb1221b81a38fdafc7d5289bfd25f8