fix keybinding input chunk shear bug
This patch fixes a problem with my previous change
to integrate keybinding support. The problem was
that we would not correctly strip keybinding sequences
if they were split across multiple chunks of input
we read off the socket. The fix is to maintain a
buffer of the current partial match if we have one
so that we can delay writing it out until after
we know if it is a complete match or not.

This involves a bunch of gross statefullness, but
I wrote a bunch of tests.

Because of the need to do this processing before we
write the output out to the pty, this kinda puts the
kibosh on my dreams of putting the keybinding engine stuff
in a background thread. Fortunately, it should be pretty
fast. I'm glad I went try hard mode on the keybinding
implementation now.

Change-Id: Ia1bf8e9b5227cf19798b16ab450e15d9748408b6