fix install script
The install script has been broken since we started
vendoring dependencies, since for some reason vendored
dependencies seem to break `cargo install --path .`.

I had been hoping to get the package in rapture and
be able to delete the install script, but the security
review is still taking a while.

Change-Id: Iafbad12d476baad349fabec6cd28dcf511bd0c66