add a default daemon config location
This patch adds a default location for the shpool
daemon config file so that users don't need to edit
their systemd unit file in order to configure shpool
with a custom keybinding. It also documents this
fact to make it easier to find.

I did not write a test for this since it would require
clobbering the shpool config for anyone who runs it.

Change-Id: I759bf83f22edb3db75dd49b1be0441851bd59bdf