fix input pty flag restoration
This patch fixes the attach subcommand to correctly
restore the pty flags if it has to use the explicit
exit trick to clobber a stuck thread. This was
causing the termianl to get mangled after a detach,
requiring a `reset` command to get normal functionality
back.

This change is pretty hard to test, so I didn't write one.

Change-Id: I2e20388ff4b0de702119dbcaf535fffa3455a1fe