update install instructions
shpool is now in rapture. This patch updates the
install instructions in the README to tell people
to use rapture rather than the old gross install
script to install shpool.

We tell people to use the unstable channel because
I want to get some people on that channel to test
stuff out before switching to recommending the stable
channel once shpool is recommended more generally.

Change-Id: I36bffa2e86be9ea4d3e384215fbb59f508a75d6f