fix bug where shpool injected a . to shell PATH
This patch fixes an issue where shpool used to inject
a `.` into the shell path. The reason it would do this
is that shpool would completely unset the PATH variable
before spawning a new shell. Bash would notice this, and
use its hardcoded default path, which for some reason includes
a `.`. This has some bad security implications. Really we want
to match the behavior of openssh completely, but the `.` part
of the bash default path is the real problem.

The fix is to hardcode the same default path that openssh
uses when it lauches a new shell. Users can also change this
value with a setting in `~/.config/shpool/config.toml`.

I considered parsing /etc/environment, but I figure it is
better to just match what openssh does.

This patch also contains some more notes in the TODO file
because I noticed some interesting stuff when reading openssh
sources and I don't want to forget about it.

FIXED=285634908

Change-Id: I78d62db6cb3ffd45f848e256ccfece3df104a232