switch exe mismatch to a warning
This patch downgrades the same exe restriction to a warning
that shows on attach. This restriction presents problems
because it forces users to immediately restart even if
they have some existing sessions. When the debian package
gets automatically upgraded under someone, this is sad
and unexpected. Versions are generaly compatible
between client and daemon processes.

Change-Id: I17d81aae8f0801f8a2acf1cb79201ea8161782b0