switch to new thread arch for session restore
This patch switches shpool to a new threading architecture
that should make the session restore feature easier to
implement. Rather than being completely dark when a user
is disconnected from a shell session, we now keep an always-on
reader thread around for each session. This thread continually
reads the output from the shell, and writes it to the terminal
spool. If a user is actually attached, then the reader thread
additionally writes the shell output directly to the socket.
We use channels to communicate with and control the reader
thread.

I was anticipating that implementing the actual output spool
itself was going to be a ton of work, but it turns out that the
already existing vt100 crate pretty much does exactly what I
wanted, so we are able to get a high quality terminal spool
without having to do any real work. The vendor churn is due to
adding that crate as a new dependency.

This change just contains threading model changes. While it does
include the configuration options for the new session restore
feature, we haven't actually implemented the two more advanced
session restore options (screen and lines). For now, we are sticking
with our old SIGWINCH tricks.

Change-Id: I02f32f7dd7d5db78fe4c5dfc60b28cc9a6af59eb