add session restore support
We did most of the work for this feature in a previous change
that refactored how we do a bunch of threading. This patch
takes advantage of that rewriting to actually add screen restore
support. By default, shpool still behaves the same way it always
has, but now it supports a new option, `session_restore_mode`,
which can either be `"simple"` (the default), `"screen"` (to restore
the entire screen, or `{ lines = n }` to restore the last n lines
on reconnect. Basically all the heavy lifting for this change
is done by the vt100 crate.

Change-Id: I431a365cf7d5aec1ccc52e52c70df9ed1e3fee6b