fix long restore chunk
This patch fixes an issue where shpool could send a session
restore chunk that exeeded the 16k limit that the attach
client is able to handle. This manifests as a session
being impossible to re-attach to. It is mostly a problem
when the user has set their `session_restore_mode` option
to `{ lines = n }` with a large n.

This patch also includes some extra logging that I put in
to try to diagnose this bug, but which I think is generally
valuable.

Change-Id: I99b5b448a0c625a09e002010bdcd7cef6fc5766d