clean up dangling reader threads better
This patch starts actually making use of the reader join
handle to join the reader thread when it has exited, but
most importaintly it starts handling the case when there is
a read error from the reader thread without the shell itself
actually exiting. In this case, the bidi stream threads would
hang around since the supervisor thread doesn't detect anything
wrong, but the reader thread would exit, so it would be impossible
to reattach or even grab the lock on the session to kill or
detach it.

I tried to write a test for this for a while, but I had
some difficulty managing to get the shell process to close
its stdout without exiting. I've observed it happening in
the wild, but replicating it in captivity proved to be annoying.
I was keeping it on ice until I could write a good test for
it, but then someone filed b/299512507 showing that this is
impacting users, so I'd like to ship it now. At the very least
the rest of the test suite still works so it shouldn't make things
any worse.

B=299512507

Change-Id: I847ebd7f2de28872f1756e745bf832a3e3e32d78