fix session restore resize trim bug
This patch fixes a bug where the session restoration data could
get mangled when a user did the following:
  1. create a new session with a wide window
  2. shrink the window down
  3. detach from shpool
  4. widen the window again
  5. reattach

Basically the problem was that this would cause shpool to only
re-draw the left side of the screen. This is because the vt100
crate has some very strange behavior in the face of resizes where
it will completely clobber lines, losing the data irrevocably,
when you resize the width down. Since it was the user's terminal
emulator rather than the vt100 crate handling all the wrapping
anyway, the fix is to just remove all width constraints on the
vt100 parser that we keep.

Change-Id: Ia5b785192df6f95c56a090157361fa802083c4c2