thread shell exit status back to attach
This patch teaches shpool to thread the exit status
of the child shell back to the attach process so that it
can exit with the same status. This seems like the right
thing to do in general since we want the attach process
to feel as much like the raw native shell as we can, but
in particular this should make it possible to exit out
of the roadwarrior reattach loop.

I bumped the pty crate version since the old version didn't
have a way to get the exit status from a child process, so
I had to add one.

BUG=295538388

Change-Id: I61b2f2784b11f74d2d2a7f54b1bd6227d22bf23b