add --ttl flag to attach
This patch adds a new --ttl flag to the attach
command, allowing users to set a finite window
of time in which a session is allowed to be up.
This is useful if you want to create sessions
with randomly generated names since you won't have
to worry too much about cleaning them up.

The changes I made for the --ttl flag exposed a
race condition in the kill mechanism and some problems
with the new exit status threading code, so in order
to get tests to pass, I had to rejigger a bunch of
stuff related to how shpool handles shell exit.
The previous exit status threading was working
in tests, but not working in the wild, so it needed
attention anyway.

B=284452708

Change-Id: Ib8f7d96e855ea6afba25453c2f661278b6e535cf