move shpool implementation into libshpool crate
This patch moves the shpool implementation into a library
crate and makes the shpool crate itself nothing more than
a one-liner to parse args and call the library's run
function.

This seems a little silly, but I'm doing this so that
we can make a google specific version of shpool with
some metrics code so that we can track our install
base and whatnot. I want to keep this repo clean for
open sourcing, and metrics code doesn't belong in an
open source project.

I had to re-vender to get the code to build, hence
the vendor churn. It is probably good to bump
deps anyway. Also, I switched to pulling the two
forked deps from crates.io rather than github
since the google import infra expects to pull
all code for 3rd party stuff from crates.io.

BUG=286063318

Change-Id: I8cb0c101fcfc66b246989145e1eb65b363b2bbdb