elide output spool in simple session restore mode
This change disables output spool processing
when in simple session restore mode so that the user only
pays for what they use.

Additionally, this patch attempts to reduce the memory overhead
of the output spool. I realized I should do that when, in
debugging another issue, I noticed that the vt100 crate eagerly fills in its
grid of terminal cells rather than lazily doing so as I had
hoped. This means that making each line have a width of
u16::MAX used up a gratitous amount of memory. After this
change it is still using way too much (10s of kilobytes per
line), but at least the situation should be somewhat improved.
The v100 crate is a big hairy ball of state, so fixing this
in a more principled way would be a lot of effort.

Change-Id: I54a3f377af98895a55088364c53e49b46bb8ed23