fix out of bounds restore bug
This patch bumps the shpool_vt100 crate version
in order to pick up a fix for a problem where
the virtual terminal could enter a corrupted
state if the cursor position was saved when
the terminal was large, then resized smaller,
and restored in the smaller mode. This was
causing the reader thread to crash.

BUG=303562211

Change-Id: I177dd0e7def1dc83cfb059f9ecb5f70b6b19bf8b