add forward_env config option
This patch adds a new config option to shpool allowing the
user to specify a list of environment variables to forward
from the shell they lauch a shpool shell from to the spawned
shell. This should make it easier to plumb environment variables
through. Right now you have to resort to gross hacks where you
smuggle the value along in one of the env vars that are forwarded.

Change-Id: I07b1e35d6b528ccd5b2815941a1cea58d1f37eff