switch to manual command line version impl
This patch switches shpool over to have it start
manually implementing version checking. As it was
before we would display the library version rather
than the version of the shpool binary. Worse than
that though is that the internal google wrapper
would just display a zero version because of the
way the toolchain works. Delegating to the wrapper
binary should make it possible to work around
the issue.

BUG=317070800

Change-Id: I53ce91dc96ee895ea0dad58b6c56d26366621150