README: update auto connect docs
The docs about setting up auto connections have
fallen a bit out of date, so this gives them a
facelift. The biggest change is significantly
simplifying the local tty based connection method
to take advantage of RemoteCommand properly,
but it also simplifies the `shpool-ssh` function
and tweaks the working for headers and tries
to do a better job of explaining why you might
want to use each approach.

Change-Id: Ie8b4b35b1156549348263e00853ba101bdf93e97