add prompt_prefix option with bash support
One issue that users complain about is that it is not
always easy to tell if you are in a shpool session or not.
The current recommendation is for them to set a prompt
hook that will annotate their prompt with the session
name, but this is asking a lot from users when shpool
is billed as a simple and easy to use session persistence
option. To address this usability issue, this patch
teaches shpool to inject the prompt hook itself.

The annotation itself is controlled by a new `prompt_prefix`
config option, which does some simple substitution of the
$SHPOOL_SESSION_NAME variable, then injects a custom prefix
in front of your prompt. The method of injection is chosen
intelligently based on if you have PROMPT_COMMAND set or not.

Initially, `prompt_prefix` will be blank by default, which
disables the feature. Once we have gathered some user feedback
I'll set a non-blank default value so that users get this prompt
annotation without any setup effort.

For the moment, this only support bash, though I'm planning
to follow up with at least a zsh hook.

I also took the opportunity to clean up the old daemon-only
--config-file option since the next version is going to break
the CLI interface due to the change in how we handle version
reporting, and we might as well get all the breakage out of
the way at once.

BUG=317069529

Change-Id: Ic852d6a975ef7de0042bf38059b5fe86888e4953