add session status to list output
This patch adds a new status column to the list output.
We don't consider new tsv columns to be breaking so long
as they are added at the end so positional access still works.
This new column should allow users to implement logic such
as automatically choosing from a pool of detached sessions
to attach to.

BUG=324020679

Change-Id: Id1c403f80bd2c96fdbf4a00d9ca74555b61da786