add new hooks API to libshpool
This patch adds a new hooks API to libshpool,
allowing the wrapping binary to register a bundle
of handlers that tell it things about the session
lifecycle. This can be useful for implementing
custom telemetry that does not belong in an open
source tool or to make a customized shpool binary
that does stuff automatically on connection drop.

This is going to cause a minor version bump right
after 0.4.0. I probably should have waited before
cutting that one.

Most of this change is actually in new test infra
since I needed to spawn the daemon as a background
thread rather than background process in order to
test this.

BUG=324072388

Change-Id: Ic7492472fd803166a9cd594cec42d94e8e5b3db9