add fish support to prompt_prefix
This patch adds fish support to the new prompt_prefix
config option. It also attempts to fix the flakyness
I've been observing in the hooks test. I think there is
a race between a session being delisted and the client
disconnect hook getting called. This is not a bug,
but it does mean the syncronization mechanism the
test was using (polling list output) is flawed. It
is better to just poll the hook records data structure
directly.

BUG=325079698

Change-Id: Iabdf49d370b1d01e4063ef74ee9df27ec146c9d6