add github ci config
This patch configures shpool to start using github
actions for CI. It is cribbed from
[wprs](https://github.com/wayland-transpositor/wprs/tree/master/.github/workflows).

I found that miri can't handle all the stuff shpool does
just yet (in particular the pipe2 syscall is not yet
supported). We should periodically check to see if sufficient
support has landed to test shpool under miri.

I had to fix a bunch of lints since this adds a new linter that
has not been run before over the codebase.