forward {x,y}pixel in term size
This patch fixes an issue where shpool would not forward
the x and y pixel values from the term size struct. These
fields are not consistently set by terminal emulators, but
some of them do set them correctly and some client programs
rely on them. We should do our part by shuttling them along.

I switched to using the pure-rust version of motd to avoid
an issue with slow startup times that was causing test
timing issues. This also makes the open source shpool work
the same way as our google internal version and means that we
have a smaller unsafe footprint, which is probably good.

I spent some effort trying to write a test for this, but
could not get wrapping the `shpool attach` subprocess in
a pty to work for whatever reason. I don't think it is worth
the effort.

Fixes #13