fix motd = "dump" mode
The current `motd = "dump"` mode will mangle the
first shell prompt that gets printed. There were
two problems, both of which are fixed in this
patch.

The first problem, was that we were only matching
on the first control code that `clear` emits. To
fix this this patch starts resolving the clear control
codes emperically rather than through terminfo queries.

The second problem was that our motd splicing logic
was buggy and would cause the next chunk after
the motd chunk to include a bunch of trailing zeros.

Fixes #15