add timeouts to session_message sends
Issue #17 describes a problem where we seem to
occasionally deadlock when the session message
handler holds the session message lock open for
a while. This patch adds some additional logging
and also adds some timeouts to the sends. The logging
should help get to the root cause if it comes up
again and the timeouts on the sends should make
it so the daemon does not get completely wedged
when this happens.