automatically reload config on changes
This patch teaches the shpool daemon to automatically
pick up changes to the config when a user writes a new
value. Right now, people need to restart the daemon
every time they make a config change, which is pretty
disruptive because it means losing sessions. Now
people can tweak config without losing sessions.

The notify crate depends on some things with some
new licences, but fortunately they are in the "notice"
and "unencumbered" classes of licences, which should
be fine to use according to google guidelines.

Fixes #20