fix bash prompt prefix injection script (#25)
This patch fixes an issue with the bash prompt prefix
injection script. It turns out PROMPT_COMMAND is actually
a list of functions to invoke rather than a single function.
This brings the bash implementation more in line with the
zsh and fish implementations and makes it work correctly
with multi-function prompt hooks.