cleanup old TODO file
I used this as a way to track tasks back when
this was nothing but a git repo with no version
control server. It has long since outlived its
usefulness. Also, both remaining entries are
worth dropping.

The heartbeat overhead is not that great and
is not causing any real problems for anyone as
far as I can tell. We can just leave it.

The setsid issue is not actually a problem
becuase the shells themselves are responsible
for calling setsid. You can verify that the
right thing is happening by running

```
ps -p $$ -o pid,sid
```

In two different shpool shells and verifying that
the pid and sid are the same in both cases. This
means that shpool shells do not inherit the same
session so long as the shell itself is well behaved.