sniff shells for prompt injection (#27)
This patch teaches shpool to discern the shell it is launching
more accurately than we were doing in the past. Previously
we would just look at the name of the shell binary, but it
turns out this does not work well when people put `exec fish`
in their .bashrc or something similar. There are actually
some good reasons for people to do this rather than use chsh
or equivalent, so we should handle this usecase.

I tried a few different approaches to this before settling
on injecting a startup sentinal to the real shell, then
scanning for it and sniffing based off the pid once the
shell is fully started up. This ought to give the rc
files enough time to get processed, and it allows
us to avoid depending on an external binary like `ps`.

To make this portable for the eventual mac port,
we use libproc to probe for the child process name
rather than directly examining /proc/<pid>/exe.

One might think that passing `--version` would work,
but this won't handle execs in .rc files.

I tested this manually and the existing per-shell prompt
prefix tests already cover it.